separate core `AbstractAnalyzer` functionalities from `JETAnalyzer` analysis
Remaining TODOs:
1. [ ] explain directory structure
2. [ ] add performance analyzer (deprecate JETTest.jl)
3. [ ] re-organize test
4. [ ] re-organize documentation

Maybe I address 2. ~ 4. in another PR.
