+++ src/abstractinterpret/abstractanalyzer.jl
@@ -1,6 +1,83 @@
-# configurations
-# --------------
+# AbstractAnalyzer
+# ================
 
+"""
+    abstract type AbstractAnalyzer <: AbstractInterpreter end
+
+An interface type of analyzers that are built on top of [JET's analyzer framework](@ref JET-Analyzer-Framework).
+
+When a new type `NewAnalyzer` implements the `AbstractAnalyzer` interface, it should be declared
+as subtype of `AbstractAnalyzer`, and is expected to the following interfaces:
+
+---
+1. `NewAnalyzer(; jetconfigs...) -> NewAnalyzer`: \\
+   Constructs new analyzer given [JET configurations](@ref) passed as `jetconfigs`.
+---
+2. `AnalyzerState(analyzer::NewAnalyzer) -> AnalyzerState`: \\
+   Returns the [`AnalyzerState`](@ref) for `analyzer::NewAnalyzer`.
+---
+3. `AbstractAnalyzer(analyzer::NewAnalyzer, state::AnalyzerState) -> NewAnalyzer`: \\
+   Constructs an new `NewAnalyzer` instance in the middle of JET's [top-level analysis](@ref toplevel)
+   or [abstract interpretation](@ref abstractinterpret), given the previous
+   `analyzer::NewAnalyzer` and [`state::AnalyzerState`](@ref AnalyzerState).
+---
+4. `ReportPass(analyzer::NewAnalyzer) -> ReportPass`: \\
+   Returns [`ReportPass`](@ref) used for `analyzer::NewAnalyzer`.
+---
+
+See also [`AnalyzerState`](@ref) and [`ReportPass`](@ref).
+
+# Example
+
+JET.jl defines its default error analyzer `JETAnalyzer <: AbstractAnalyzer` as the following
+(modified a bit for the sake of simplicity):
+```julia
+# the default error analyzer for JET.jl
+struct JETAnalyzer{RP<:ReportPass} <: AbstractAnalyzer
+    report_pass::RP
+    state::AnalyzerState
+end
+
+# AbstractAnalyzer API requirements
+
+function JETAnalyzer(;
+    report_pass::ReportPass = BasicPass(),
+    jetconfigs...)
+    return JETAnalyzer(report_pass,
+                       AnalyzerState(; jetconfigs...),
+                       )
+end
+AnalyzerState(analyzer::JETAnalyzer) =
+    return analyzer.state
+AbstractAnalyzer(analyzer::JETAnalyzer, state::AnalyzerState) =
+    return JETAnalyzer(ReportPass(analyzer), state)
+ReportPass(analyzer::JETAnalyzer) =
+    return analyzer.report_pass
+```
+"""
+abstract type AbstractAnalyzer <: AbstractInterpreter end
+
+Base.show(io::IO, analyzer::AbstractAnalyzer) = print(io, "AbstractAnalyzer")
+Base.show(io::IO, ::MIME"application/prs.juno.inline", analyzer::AbstractAnalyzer) =
+    return analyzer
+
+# interface 1
+# -----------
+# 1. `NewAnalyzer(; jetconfigs...) -> NewAnalyzer`
+
+function (::Type{Analyzer})(; jetconfigs...) where Analyzer<:AbstractAnalyzer
+    error("""
+    missing `$AbstractAnalyzer` API:
+    `$Analyzer` is required to implement the `$Analyzer(; jetconfigs...) -> $Analyzer` interface.
+    See the documentation of `$AbstractAnalyzer`.
+    """)
+end
+
+# interface 2
+# -----------
+# 2. `AnalyzerState(analyzer::NewAnalyzer) -> AnalyzerState`
+
+# TODO move this to `JETAnalyzer`
 """
 Configurations for abstract interpretation performed by JET.
 These configurations will be active for all the entries.
@@ -90,49 +167,6 @@ struct JETAnalysisParams
     end
 end
 
-# define these functions just to make them able to accept other JET configrations
-JETInferenceParams(; ipo_constant_propagation::Bool        = true,
-                     aggressive_constant_propagation::Bool = false,
-                     unoptimize_throw_blocks::Bool         = true,
-                     max_methods::Int                      = 3,
-                     union_splitting::Int                  = 4,
-                     apply_union_enum::Int                 = 8,
-                     tupletype_depth::Int                  = 3,
-                     tuple_splat::Int                      = 32,
-                     __jetconfigs...) =
-    return InferenceParams(; ipo_constant_propagation,
-                             aggressive_constant_propagation,
-                             unoptimize_throw_blocks,
-                             max_methods,
-                             union_splitting,
-                             apply_union_enum,
-                             tupletype_depth,
-                             tuple_splat,
-                             )
-JETOptimizationParams(; inlining::Bool                = inlining_enabled(),
-                        inline_cost_threshold::Int    = 100,
-                        inline_nonleaf_penalty::Int   = 1000,
-                        inline_tupleret_bonus::Int    = 250,
-                        inline_error_path_cost::Int   = 20,
-                        max_methods::Int              = 3,
-                        tuple_splat::Int              = 32,
-                        union_splitting::Int          = 4,
-                        unoptimize_throw_blocks::Bool = true,
-                        __jetconfigs...) =
-    return OptimizationParams(; inlining,
-                                inline_cost_threshold,
-                                inline_nonleaf_penalty,
-                                inline_tupleret_bonus,
-                                inline_error_path_cost,
-                                max_methods,
-                                tuple_splat,
-                                union_splitting,
-                                unoptimize_throw_blocks,
-                                )
-# # assert here that they create same objects as the original constructors
-@assert JETInferenceParams() == InferenceParams()
-@assert JETOptimizationParams() == OptimizationParams()
-
 const LOGGER_LEVEL_KEY     = :JET_LOGGER_LEVEL
 const INFO_LOGGER_LEVEL    = 0
 const DEBUG_LOGGER_LEVEL   = 1
@@ -211,6 +245,32 @@ struct JETLogger
     end
 end
 
+"""
+    mutable struct AnalyzerState
+        ...
+    end
+
+The mutable object that holds various states that are consumed by all [`AbstractAnalyzer`](@ref)s.
+
+---
+
+    AnalyzerState(analyzer::AbstractAnalyzer) -> AnalyzerState
+
+If `NewAnalyzer` implements the `AbstractAnalyzer` interface, `NewAnalyzer` should implement
+this `AnalyzerState(analyzer::NewAnalyzer) -> AnalyzerState` interface.
+
+A new `AnalyzerState` is supposed to be constructed using [JET configurations](@ref) passed
+as keyword arguments `jetconfigs` of the [`NewAnalyzer(; jetconfigs...)`](@ref AbstractAnalyzer)
+constructor, and the constructed `AnalyzerState` is usually kept within `NewAnalyzer` itself:
+```julia
+function NewAnalyzer(; jetconfigs...)
+    ...
+    state = AnalyzerState(; jetconfigs...)
+    return NewAnalyzer(..., state)
+end
+AnalyzerState(analyzer::NewAnalyzer) = analyzer.state
+```
+"""
 mutable struct AnalyzerState
     ## native ##
 
@@ -250,49 +310,7 @@ mutable struct AnalyzerState
     depth::Int
 end
 
-"""
-    mutable struct AnalyzerState
-        ...
-    end
-
-The mutable object that holds various states that are consumed by all [`AbstractAnalyzer`](@ref)s.
-
----
-
-    AnalyzerState(analyzer::AbstractAnalyzer) -> AnalyzerState
-
-If `NewAnalyzer` implements the `AbstractAnalyzer` interface, `NewAnalyzer` should implement
-this `AnalyzerState(analyzer::NewAnalyzer) -> AnalyzerState` interface.
-
-A new `AnalyzerState` is supposed to be constructed using [JET configurations](@ref) passed
-as keyword arguments `jetconfigs` of the [`NewAnalyzer(; jetconfigs...)`](@ref AbstractAnalyzer)
-constructor, and the constructed `AnalyzerState` is usually kept within `NewAnalyzer` itself:
-```julia
-function NewAnalyzer(; jetconfigs...)
-    ...
-    state = AnalyzerState(; jetconfigs...)
-    return NewAnalyzer(..., state)
-end
-AnalyzerState(analyzer::NewAnalyzer) = analyzer.state
-```
-"""
-function AnalyzerState(analyzer::Analyzer) where Analyzer<:AbstractAnalyzer
-    error("""
-    missing `$AbstractAnalyzer` API:
-    `$Analyzer` is required to implement the `$AnalyzerState(analyzer::$Analyzer) -> $AnalyzerState` interface.
-    See the documentation of `$AbstractAnalyzer` and `$AnalyzerState`.
-    """)
-end
-
-function AbstractAnalyzer(analyzer::Analyzer, state::AnalyzerState) where Analyzer<:AbstractAnalyzer
-    error("""
-    missing `$AbstractAnalyzer` API:
-    `$Analyzer` is required to implement the `$AbstractAnalyzer(analyzer::$Analyzer, state::$AnalyzerState) -> $Analyzer` interface.
-    See the documentation of `$AbstractAnalyzer`.
-    """)
-end
-
-# define getter/setter methods
+# define shortcut getter/setter methods for `AbstractAnalyzer`s
 for fld in fieldnames(AnalyzerState)
     fld === :cache_key && continue # will be defined later (in order to take in `ReportPass` hash)
     getter = Symbol("get_", fld)
@@ -336,6 +354,83 @@ const _CONCRETIZED  = BitVector()
 const _TOPLEVELMOD  = __toplevelmod__
 const _GLOBAL_SLOTS = Dict{Int,Symbol}()
 
+# wrappers of `InferenceParams` and `OptimizationParams` that can accept JET configrations
+JETInferenceParams(; ipo_constant_propagation::Bool        = true,
+                     aggressive_constant_propagation::Bool = false,
+                     unoptimize_throw_blocks::Bool         = true,
+                     max_methods::Int                      = 3,
+                     union_splitting::Int                  = 4,
+                     apply_union_enum::Int                 = 8,
+                     tupletype_depth::Int                  = 3,
+                     tuple_splat::Int                      = 32,
+                     __jetconfigs...) =
+    return InferenceParams(; ipo_constant_propagation,
+                             aggressive_constant_propagation,
+                             unoptimize_throw_blocks,
+                             max_methods,
+                             union_splitting,
+                             apply_union_enum,
+                             tupletype_depth,
+                             tuple_splat,
+                             )
+JETOptimizationParams(; inlining::Bool                = inlining_enabled(),
+                        inline_cost_threshold::Int    = 100,
+                        inline_nonleaf_penalty::Int   = 1000,
+                        inline_tupleret_bonus::Int    = 250,
+                        inline_error_path_cost::Int   = 20,
+                        max_methods::Int              = 3,
+                        tuple_splat::Int              = 32,
+                        union_splitting::Int          = 4,
+                        unoptimize_throw_blocks::Bool = true,
+                        __jetconfigs...) =
+    return OptimizationParams(; inlining,
+                                inline_cost_threshold,
+                                inline_nonleaf_penalty,
+                                inline_tupleret_bonus,
+                                inline_error_path_cost,
+                                max_methods,
+                                tuple_splat,
+                                union_splitting,
+                                unoptimize_throw_blocks,
+                                )
+# # assert here that they create same objects as the original constructors
+@assert JETInferenceParams() == InferenceParams()
+@assert JETOptimizationParams() == OptimizationParams()
+
+function AnalyzerState(analyzer::Analyzer) where Analyzer<:AbstractAnalyzer
+    error("""
+    missing `$AbstractAnalyzer` API:
+    `$Analyzer` is required to implement the `$AnalyzerState(analyzer::$Analyzer) -> $AnalyzerState` interface.
+    See the documentation of `$AbstractAnalyzer` and `$AnalyzerState`.
+    """)
+end
+
+# interface 3
+# -----------
+# 3. `AbstractAnalyzer(analyzer::NewAnalyzer, state::AnalyzerState) -> NewAnalyzer`
+
+function AbstractAnalyzer(analyzer::Analyzer, state::AnalyzerState) where Analyzer<:AbstractAnalyzer
+    error("""
+    missing `$AbstractAnalyzer` API:
+    `$Analyzer` is required to implement the `$AbstractAnalyzer(analyzer::$Analyzer, state::$AnalyzerState) -> $Analyzer` interface.
+    See the documentation of `$AbstractAnalyzer`.
+    """)
+end
+
+# constructor for additional JET analysis in the middle of parent (non top-level) abstractinterpret
+function AbstractAnalyzer(analyzer::T) where {T<:AbstractAnalyzer}
+    newstate = AnalyzerState(get_world_counter(analyzer);
+                             analysis_params = JETAnalysisParams(analyzer),
+                             inf_params      = InferenceParams(analyzer),
+                             opt_params      = OptimizationParams(analyzer),
+                             logger          = JETLogger(analyzer),
+                             depth           = get_depth(analyzer),
+                             )
+    newanalyzer = AbstractAnalyzer(analyzer, newstate)
+    maybe_initialize_caches!(newanalyzer)
+    return newanalyzer
+end
+
 # constructor for sequential toplevel JET analysis
 function AbstractAnalyzer(analyzer::T, concretized, toplevelmod) where {T<:AbstractAnalyzer}
     newstate = AnalyzerState(# update world age to take in newly added methods defined
@@ -353,37 +448,89 @@ function AbstractAnalyzer(analyzer::T, concretized, toplevelmod) where {T<:Abstr
     return newanalyzer
 end
 
-# constructor to do additional JET analysis in the middle of parent (non-toplevel) interpretation
-function AbstractAnalyzer(analyzer::T) where {T<:AbstractAnalyzer}
-    newstate = AnalyzerState(get_world_counter(analyzer);
-                             analysis_params = JETAnalysisParams(analyzer),
-                             inf_params      = InferenceParams(analyzer),
-                             opt_params      = OptimizationParams(analyzer),
-                             logger          = JETLogger(analyzer),
-                             depth           = get_depth(analyzer),
-                             )
-    newanalyzer = AbstractAnalyzer(analyzer, newstate)
-    maybe_initialize_caches!(newanalyzer)
-    return newanalyzer
+# interface 4
+# -----------
+# 4. `ReportPass(analyzer::NewAnalyzer) -> ReportPass`
+
+"""
+    abstract type ReportPass end
+
+An interface type that represents `AbstractAnalyzer`'s report pass.
+`analyzer::AbstractAnalyzer` injects report passes using the `(::ReportPass)(::Type{InferenceErrorReport}, ::AbstractAnalyzer, state, ...)`
+interface, which provides a flexible and efficient layer to configure the analysis done by `AbstractAnalyzer`.
+
+---
+
+    ReportPass(analyzer::AbstractAnalyzer) -> ReportPass
+
+If `NewAnalyzer` implements the `AbstractAnalyzer` interface, `NewAnalyzer` should implement
+this `ReportPass(analyzer::NewAnalyzer) -> ReportPass` interface.
+
+`ReportPass` allows `NewAnalyzer` to provide a very flexible configuration layer for `NewAnalyzer`'s analysis;
+an user can define their own `ReportPass` to control how `NewAnalyzer` collects report errors
+while still using the analysis routine implemented by `NewAnalyzer`.
+
+# Example
+
+For example, [`JETAnalyzer`](@ref) accepts a custom `ReportPass` passed as [JET configurations](@ref)
+(see the example documentation of [`AbstractAnalyzer`](@ref) for details).
+And we can setup a custom report pass `IgnoreAllExceptGlobalUndefVar`, that ignores all the
+reports that are otherwise collected by `JETAnalyzer` except `GlobalUndefVarErrorReport`:
+```julia
+# custom report pass that ignores all the reports except `GlobalUndefVarErrorReport`
+struct IgnoreAllExceptGlobalUndefVar <: ReportPass end
+
+# ignores all the reports analyzed by `JETAnalyzer`
+(::IgnoreAllExceptGlobalUndefVar)(::Type{<:InferenceErrorReport}, @nospecialize(_...)) = return
+
+# bypass to `BasicPass` to collect `GlobalUndefVarErrorReport`
+function (::IgnoreAllExceptGlobalUndefVar)(::Type{GlobalUndefVarErrorReport}, @nospecialize(args...))
+    BasicPass()(GlobalUndefVarErrorReport, args...)
 end
 
-Base.show(io::IO, analyzer::AbstractAnalyzer) = print(io, "AbstractAnalyzer")
-Base.show(io::IO, ::MIME"application/prs.juno.inline", analyzer::AbstractAnalyzer) =
-    return analyzer
-Base.show(io::IO, ::MIME"application/prs.juno.inline", analyzer::NativeInterpreter) =
-    return analyzer
-
-# XXX: this should be upstreamed
-function Base.show(io::IO, frame::InferenceState)
-    print(io, "InfernceState for ")
-    show(io, frame.linfo)
-    print(io, " at pc ", frame.currpc, '/', length(frame.src.code))
+no_method_error()    = 1 + "1"
+undef_global_error() = undefvar
+report_call(; report_pass=IgnoreAllExceptGlobalUndefVar()) do
+    if rand(Bool)
+        return no_method_error()    # "no matching method found" error report won't be reported here
+    else
+        return undef_global_error() # "variable undefvar is not defined" error report will be reported
+    end
 end
-Base.show(io::IO, ::MIME"application/prs.juno.inline", frame::InferenceState) =
-    return frame
+```
+"""
+abstract type ReportPass end
+
+function ReportPass(::Analyzer) where Analyzer<:AbstractAnalyzer
+    error("""
+    missing `$AbstractAnalyzer` API:
+    `$Analyzer` is required to implement the `$ReportPass(analyzer::$Analyzer) -> $ReportPass` interface.
+    See the documentation of `$AbstractAnalyzer` and `$ReportPass`.
+    """)
+end
+# some specific reports are necessary to be collected during `AbstractAnalyzer`'s core routine
+# and ignore them by default (analyzer can opt-in to collect them by overloading this with
+# their own report pass)
+# otherwise, it means malformed report pass call, and we should inform users of it
+function (rp::ReportPass)(T, @nospecialize(args...))
+    if !(T === NativeRemark ||
+         T === InvalidConstantRedefinition ||
+         T === InvalidConstantDeclaration)
+        throw(MethodError(rp, (T, args...)))
+    end
+end
+
+# TODO separate this from `AbstractAnalyzer`
+# default constructor to create a report from abstract interpretation routine
+function (T::Type{<:InferenceErrorReport})(analyzer::AbstractAnalyzer, state, @nospecialize(spec_args...))
+    vf = get_virtual_frame(state)
+    msg = get_msg(T, analyzer, state, spec_args...)
+    return T([vf], msg, vf.sig, spec_args...)
+end
 
 # InferenceResult
-# ---------------
+# ===============
+# define how AbstractAnalyzer manages `InferenceResult`
 
 const Reports = Vector{InferenceErrorReport}
 const CachedReports = Vector{InferenceErrorReportCache}
@@ -461,8 +608,9 @@ add_caller_cache!(analyzer::AbstractAnalyzer, report::InferenceErrorReport) =
 add_caller_cache!(analyzer::AbstractAnalyzer, reports::Vector{InferenceErrorReport}) =
     return append!(get_caller_cache(analyzer), reports)
 
-# AbstractInterpreter API
-# -----------------------
+# AbstractInterpreter
+# ===================
+# provide default implementations for the API requirements
 
 CC.InferenceParams(analyzer::AbstractAnalyzer)    = InferenceParams(get_native(analyzer))
 CC.OptimizationParams(analyzer::AbstractAnalyzer) = OptimizationParams(get_native(analyzer))
@@ -472,7 +620,19 @@ CC.get_world_counter(analyzer::AbstractAnalyzer)  = get_world_counter(get_native
 CC.lock_mi_inference(::AbstractAnalyzer, ::MethodInstance) = nothing
 CC.unlock_mi_inference(::AbstractAnalyzer, ::MethodInstance) = nothing
 
-CC.add_remark!(analyzer::AbstractAnalyzer, sv, s) = ReportPass(analyzer)(NativeRemark, analyzer, sv, s)
+"""
+    NativeRemark <: InferenceErrorReport
+
+This special `InferenceErrorReport` wraps remarks by `NativeInterpreter`.
+"remarks" are information that Julia's native compiler emits about how its type inference goes,
+and those remarks are less interesting in term of "error checking", so currently any of JET's
+pre-defined report passes doesn't make any use of `NativeRemark`.
+"""
+@reportdef struct NativeRemark <: InferenceErrorReport
+    s::String
+end
+get_msg(::Type{NativeRemark}, analyzer::AbstractAnalyzer, sv, s) = s
+CC.add_remark!(analyzer::AbstractAnalyzer, sv, s) = ReportPass(analyzer)(NativeRemark, analyzer, sv, s) # ignored by default
 
 CC.may_optimize(analyzer::AbstractAnalyzer)      = true
 CC.may_compress(analyzer::AbstractAnalyzer)      = false
@@ -519,7 +679,8 @@ end
 end # @static if isdefined(Compiler, :is_stmt_inline)
 
 # AbstractAnalyzer
-# ----------------
+# ================
+# AbstractAnalyzer specific APIs
 
 JETAnalysisParams(analyzer::AbstractAnalyzer) = get_analysis_params(analyzer)
 
@@ -542,10 +703,6 @@ function maybe_initialize_caches!(analyzer::AbstractAnalyzer)
         JET_CACHE[cache_key] = IdDict{MethodInstance,CodeInstance}()
     end
 end
-
-# check if we're in a toplevel module
-istoplevel(sv::State)             = istoplevel(sv.linfo)
-istoplevel(linfo::MethodInstance) = isa(linfo.def, Module)
 
 is_global_slot(analyzer::AbstractAnalyzer, slot::Int)   = slot in keys(get_global_slots(analyzer))
 is_global_slot(analyzer::AbstractAnalyzer, slot::Slot)  = is_global_slot(analyzer, slot_id(slot))
@@ -571,65 +728,40 @@ JETLogger(analyzer::AbstractAnalyzer) = get_logger(analyzer)
 end
 
 """
-    NativeRemark <: InferenceErrorReport
+    aggregation_policy(analyzer::AbstractAnalyzer)
 
-This special `InferenceErrorReport` wraps remarks by `NativeInterpreter`.
-"remarks" are information that Julia's native compiler emits about how its type inference goes,
-and those remarks are less interesting in term of "error checking", so currently any of JET's
-pre-defined report passes doesn't make any use of `NativeRemark`.
+Defines how `analyzer` aggregates [`InferenceErrorReport`](@ref)s.
+Defaults to `default_aggregation_policy`.
+
+---
+
+    default_aggregation_policy(report::InferenceErrorReport) -> DefaultReportIdentity
+
+Returns the default identity of `report::InferenceErrorReport`, where `DefaultReportIdentity`
+aggregates reports based on "error location" of each `report`.
+`DefaultReportIdentity` aggregates `InferenceErrorReport`s aggressively in a sense that it
+ignores the identity of error point's `MethodInstance`, under the assumption that errors are
+identical as far as they're collected at the same file and line.
 """
-@reportdef struct NativeRemark <: InferenceErrorReport
-    s::String
+aggregation_policy(::AbstractAnalyzer) = default_aggregation_policy
+function default_aggregation_policy(@nospecialize(report::InferenceErrorReport))
+    return DefaultReportIdentity(
+        typeof(report),
+        report.sig,
+        # VirtualFrameNoLinfo(first(report.vst)),
+        VirtualFrameNoLinfo(last(report.vst)),
+        )
 end
-get_msg(::Type{NativeRemark}, analyzer::AbstractAnalyzer, sv, s) = s
-
-# ignores `NativeRemark`s by default
-(::SoundBasicPass)(::Type{NativeRemark}, analyzer::AbstractAnalyzer, sv::InferenceState, s) = return
-
-# JETAnalyzer
-# -----------
-
-"""
-    struct JETAnalyzer <: AbstractAnalyzer
-
-JET.jl's default error analyzer.
-"""
-struct JETAnalyzer{RP<:ReportPass} <: AbstractAnalyzer
-    report_pass::RP
-    state::AnalyzerState
+@withmixedhash struct VirtualFrameNoLinfo
+    file::Symbol
+    line::Int
+    sig::Vector{Any}
+    # linfo::MethodInstance # ignore the idenity of `MethodInstace`
+    VirtualFrameNoLinfo(vf::VirtualFrame) = new(vf.file, vf.line, vf.sig)
 end
-
-# AbstractAnalyzer API requirements
-
-# NOTE `@aggressive_constprop` here makes sure `mode` to be propagated as constant
-@aggressive_constprop @jetconfigurable function JETAnalyzer(;
-    report_pass::Union{Nothing,T} = nothing,
-    mode::Symbol                  = :basic,
-    # default `InferenceParams` tuning
-    aggressive_constant_propagation::Bool = true,
-    unoptimize_throw_blocks::Bool         = false,
-    # default `OptimizationParams` tuning
-    inlining::Bool = false,
-    jetconfigs...) where {T<:ReportPass}
-    if isnothing(report_pass)
-        # if `report_pass` isn't passed explicitly, here we configure it according to `mode`
-        report_pass = mode === :basic ? BasicPass() :
-                      mode === :sound ? SoundPass() :
-                      throw(ArgumentError("`mode` configuration should be either of `:basic` or `:sound`"))
-    end
-    # NOTE we always disable inlining, because:
-    # - our current strategy to find undefined local variables and uncaught `throw` calls assumes un-inlined frames
-    # - the cost for inlining isn't necessary for JETAnalyzer
-    @assert !inlining "inlining should be disabled for JETAnalyzer"
-    state = AnalyzerState(; aggressive_constant_propagation,
-                            unoptimize_throw_blocks,
-                            inlining,
-                            jetconfigs...)
-    return JETAnalyzer(report_pass, state)
+@withmixedhash struct DefaultReportIdentity
+    T::Type{<:InferenceErrorReport}
+    sig::Vector{Any}
+    # entry_frame::VirtualFrameNoLinfo
+    error_frame::VirtualFrameNoLinfo
 end
-AnalyzerState(analyzer::JETAnalyzer) =
-    return analyzer.state
-AbstractAnalyzer(analyzer::JETAnalyzer, state::AnalyzerState) =
-    return JETAnalyzer(ReportPass(analyzer), state)
-ReportPass(analyzer::JETAnalyzer) =
-    return analyzer.report_pass
