+++ src/ui/print.jl
@@ -331,8 +331,3 @@ function print_error_report(io, report::InferenceErrorReport)
                     bold = true,
                     )
 end
-# those don't need explicit signatures
-print_error_report(io, report::NoFieldErrorReport)       = printlnstyled(io, "│ ", report.msg; color = ERROR_COLOR)
-print_error_report(io, report::LocalUndefVarErrorReport) = printlnstyled(io, "│ ", report.msg; color = ERROR_COLOR)
-print_error_report(io, report::DivideErrorReport)        = printlnstyled(io, "│ ", report.msg; color = ERROR_COLOR)
-print_error_report(io, report::SeriousExceptionReport)   = printlnstyled(io, "│ ", report.msg; color = ERROR_COLOR)
